refactor: create folder contain all file type schema
